clients/js: re-org sui
