+++ clients/js/sui/index.ts
@@ -1,4 +1,7 @@
 export * from "./MoveToml";
+export * from "./buildCoin";
+export * from "./log";
 export * from "./publish";
+export * from "./submit";
 export * from "./types";
 export * from "./utils";
