+++ clients/js/sui/utils.ts
@@ -1,5 +1,3 @@
-import { CHAIN_ID_SUI } from "@certusone/wormhole-sdk";
-import { CHAIN_ID_TO_NAME } from "@certusone/wormhole-sdk/lib/cjs/utils/consts";
 import {
   Connection,
   Ed25519Keypair,
@@ -13,116 +11,13 @@ import {
   getPublishedObjectChanges,
   normalizeSuiAddress,
 } from "@mysten/sui.js";
-import { CONTRACTS } from "../consts";
 import { NETWORKS } from "../networks";
 import { Network } from "../utils";
-import { Payload, VAA, impossible, parse, serialiseVAA } from "../vaa";
+import { Payload, VAA, parse, serialiseVAA } from "../vaa";
 import { SuiRpcValidationError } from "./error";
 import { SuiCreateEvent, SuiPublishEvent } from "./types";
 
 const UPGRADE_CAP_TYPE = "0x2::package::UpgradeCap";
-
-export const execute_sui = async (
-  payload: Payload,
-  vaa: Buffer,
-  network: Network,
-  rpc?: string,
-  privateKey?: string
-) => {
-  const chain = CHAIN_ID_TO_NAME[CHAIN_ID_SUI];
-  const provider = getProvider(network, rpc);
-  const signer = getSigner(provider, network, privateKey);
-
-  switch (payload.module) {
-    case "Core": {
-      const coreObjectId = CONTRACTS[network][chain].core;
-      if (!coreObjectId) {
-        throw Error("Core bridge object ID is undefined");
-      }
-      const corePackageId = await getPackageId(provider, coreObjectId);
-      switch (payload.type) {
-        case "GuardianSetUpgrade": {
-          console.log("Submitting new guardian set");
-          const tx = new TransactionBlock();
-          if (network === "DEVNET") {
-            // Avoid Error checking transaction input objects: GasBudgetTooHigh { gas_budget: 50000000000, max_budget: 10000000000 }
-            tx.setGasBudget(10000000000);
-          }
-          tx.moveCall({
-            target: `${corePackageId}::wormhole::update_guardian_set`,
-            arguments: [
-              tx.object(coreObjectId),
-              tx.pure([...vaa]),
-              tx.object(SUI_CLOCK_OBJECT_ID),
-            ],
-          });
-          const result = await executeTransactionBlock(signer, tx);
-          console.log(JSON.stringify(result));
-          break;
-        }
-        case "ContractUpgrade":
-          throw new Error("ContractUpgrade not supported on Sui");
-        case "RecoverChainId":
-          throw new Error("RecoverChainId not supported on Sui");
-        default:
-          impossible(payload);
-      }
-      break;
-    }
-    case "NFTBridge": {
-      throw new Error("NFT bridge not supported on Sui");
-    }
-    case "TokenBridge": {
-      const coreBridgeStateObjectId = CONTRACTS[network][chain].core;
-      if (!coreBridgeStateObjectId) {
-        throw Error("Core bridge object ID is undefined");
-      }
-
-      const tokenBridgeStateObjectId = CONTRACTS[network][chain].token_bridge;
-      if (!tokenBridgeStateObjectId) {
-        throw Error("Token bridge object ID is undefined");
-      }
-
-      switch (payload.type) {
-        case "ContractUpgrade":
-          throw new Error("ContractUpgrade not supported on Sui");
-        case "RecoverChainId":
-          throw new Error("RecoverChainId not supported on Sui");
-        case "RegisterChain": {
-          console.log("Registering chain");
-          if (payload.module !== "TokenBridge") {
-            throw new Error("NFT bridge not supported on Sui");
-          }
-
-          const signer = getSigner(provider, network, privateKey);
-          const tx = await registerChain(
-            provider,
-            network,
-            vaa,
-            coreBridgeStateObjectId,
-            tokenBridgeStateObjectId
-          );
-          const res = await executeTransactionBlock(signer, tx);
-          console.log(JSON.stringify(res));
-          break;
-        }
-        case "AttestMeta":
-          throw new Error("AttestMeta not supported on Sui");
-        case "Transfer":
-          throw new Error("Transfer not supported on Sui");
-        case "TransferWithPayload":
-          throw Error("Can't complete payload 3 transfer from CLI");
-        default:
-          impossible(payload);
-          break;
-      }
-
-      break;
-    }
-    default:
-      impossible(payload);
-  }
-};
 
 export const executeTransactionBlock = async (
   signer: RawSigner,
